Refactor `JupiterTestEngineTests`
Issue: #2816
